moved shared python test file to be consistent
