+++ tests/python/test_cloudrun.py
@@ -21,7 +21,7 @@ import google.cloud.logging
 from google.cloud.logging_v2.resource import Resource
 
 from ..common.common import Common
-from ..common.python import CommonPython
+from .python import CommonPython
 
 
 class TestCloudRun(Common, CommonPython, unittest.TestCase):
